requirements.txt
accessify==0.3.1
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
boto3==1.23.10
botocore==1.26.10
certifi==2022.5.18.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
colorama==0.4.4
commonmark==0.9.1
cryptography==37.0.2
decorator==5.1.1
Django==3.2.12
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
executing==0.8.3
gunicorn==20.1.0
h11==0.13.0
idna==2.8
ipython==8.1.1
jedi==0.18.1
jmespath==1.0.1
matplotlib-inline==0.1.3
outcome==1.1.0
parso==0.8.3
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pypi-version==0.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.1
requests==2.28.1
rich==12.4.4
s3transfer==0.5.2
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.2.0
toml==0.10.2
tqdm==4.62.3
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.11
watchdog==2.1.8
wcwidth==0.2.5
whitenoise==6.2.0
wsproto==1.1.0
